Reorganise runner code
